Extract merger
